'/users' and '/auth' refactoring + unified docstrings + single response schema.

+++ api_server/users/services/__init__.py
@@ -2,48 +2,96 @@ from typing import Union
 from uuid import UUID
 import abc
 
-from fastapi import status
+from fastapi import Depends, status
 
+from fastapi_jwt_auth import AuthJWT
 from passlib.hash import argon2
 from sqlalchemy import update
 from sqlalchemy.exc import NoResultFound
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.future import select
 
+from db import get_session
 from users.models import User
-from users.schemas import UserInputSchema
+from users.schemas import UserInputSchema, UserUpdateSchema
 from users.utils.exceptions import UserNotFoundError
+from users.utils.jwt import jwt_user_validator
 from utils.logging import setup_logging
 
 
 class AbstractUserService(metaclass=abc.ABCMeta):
 
-    def __init__(self, session: AsyncSession) -> None:
+    def __init__(
+        self,
+        session: AsyncSession = Depends(get_session),
+        Authorize: AuthJWT = Depends(),
+    ) -> None:
         self._log = setup_logging(self.__class__.__name__)
         self.session = session
+        self.Authorize = Authorize
 
-    async def get_users(self) -> None:
-        """Return list of User objects from the db."""
+    async def get_users(self) -> list[User]:
+        """Get User objects from database.
+
+        Returns:
+        list of User objects.
+        """
         return await self._get_users()
 
-    async def get_user_by_id(self, id_: str) -> None:
-        """Return single User object from the db filtered by 'id'."""
+    async def get_user_by_id(self, id_: str) -> User:
+        """Get User object from database filtered by id.
+
+        Args:
+            id_: UUID of user.
+
+        Returns:
+        single User object filtered by id.
+        """
         return await self._get_user_by_id(id_)
 
-    async def add_user(self, user: UserInputSchema) -> None:
-        """Add User object to the db."""
+    async def add_user(self, user: UserInputSchema) -> User:
+        """Add User object to the database.
+
+        Args:
+            user: UserInputSchema object.
+
+        Returns:
+        newly created User object.
+        """
         return await self._add_user(user)
 
-    async def update_user(self, id_: str, user: UserInputSchema) -> None:
-        """Update User object in the db."""
+    async def update_user(self, id_: str, user: UserUpdateSchema) -> User:
+        """Updates User object in the database.
+
+        Args:
+            id_: UUID of user.
+            user: UserUpdateSchema object.
+
+        Returns:
+        updated User object.
+        """
         return await self._update_user(id_, user)
 
     async def delete_user(self, id_: str) -> None:
-        """Delete User object from the db."""
+        """Delete User object from the database.
+
+        Args:
+            id_: UUID of user.
+
+        Returns:
+        Nothing.
+        """
         return await self._delete_user(id_)
 
-    async def get_user_by_username(self, username: str) -> None:
-        """Return User object filtered by username."""
+    async def get_user_by_username(self, username: str) -> User:
+        """Get User object from database filtered by username.
+
+        Args:
+            username: string of User's username.
+
+        Returns:
+        Single User object filtered by username.
+        """
         return await self._get_user_by_username(username)
 
     @abc.abstractclassmethod
@@ -59,7 +107,7 @@ class AbstractUserService(metaclass=abc.ABCMeta):
         pass
 
     @abc.abstractclassmethod
-    async def _update_user(self, id_: str, user: UserInputSchema) -> None:
+    async def _update_user(self, id_: str, user: UserUpdateSchema) -> None:
         pass
 
     @abc.abstractclassmethod
@@ -108,23 +156,29 @@ class UserService(AbstractUserService):
         await self.session.refresh(user)
         return user
 
-    async def _update_user(self, id_: str, user: UserInputSchema) -> None:
+    async def _update_user(self, id_: str, user: UserUpdateSchema) -> None:
+        self.Authorize.jwt_required()
+        jwt_subject = self.Authorize.get_jwt_subject()
         user_exists = await self._select_user(column='id', value=id_)
         if user_exists:
-            # Updating user.
-            await self.session.execute(update(User).where(User.id == id_).values(**user.dict()))
-            await self.session.commit()
-            # Return updated user.
-            return await self._get_user_by_id(id_=id_)
+            if jwt_user_validator(jwt_subject=jwt_subject, username=user_exists.username):
+                # Updating user.
+                await self.session.execute(update(User).where(User.id == id_).values(**user.dict()))
+                await self.session.commit()
+                # Return updated user.
+                return await self._get_user_by_id(id_=id_)
 
     async def _delete_user(self, id_) -> None:
+        self.Authorize.jwt_required()
+        jwt_subject = self.Authorize.get_jwt_subject()
         user_exists = await self._select_user(column='id', value=id_)
         if user_exists:
-            # Deleting user.
-            user = await self._get_user_by_id(id_=id_)
-            await self.session.delete(user)
-            await self.session.commit()
-            self._log.debug(f'''User with id: "{id_}" deleted.''')
+            if jwt_user_validator(jwt_subject=jwt_subject, username=user_exists.username):
+                # Deleting user.
+                user = await self._get_user_by_id(id_=id_)
+                await self.session.delete(user)
+                await self.session.commit()
+                self._log.debug(f'''User with id: "{id_}" deleted.''')
 
     async def _hash_password(self, password: str) -> str:
         """Return password hashed with argon2 algorithm."""
